feat(filesystem): add new project module filesystem management to save file like s3 aws

+++ filesystem/src/main/java/com/tmdad/filesystem/service/StorageService.java
@@ -1,4 +1,4 @@
-package com.tmda.chatapp.service;
+package com.tmdad.filesystem.service;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,29 +8,46 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.FileSystemUtils;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 @Service
 public class StorageService {
 
-    Logger log = LoggerFactory.getLogger(this.getClass().getName());
-
-//   TODO PATH FILES
-    private final Path rootLocation = Paths.get("upload-dir");
+    private static final Logger logger = LoggerFactory.getLogger(StorageService.class.getName());
+    private final Path rootLocation = Paths.get("../files");
 
     public void store(MultipartFile file) {
+        logger.info("Call store with file name: {}", file.getName());
         try {
-            Files.copy(file.getInputStream(), this.rootLocation.resolve(file.getOriginalFilename()));
+            String date = new SimpleDateFormat("yyyy-MM-dd").format(new Date());
+            File f = new File(this.rootLocation.resolve(file.getOriginalFilename()).toString());
+            if(f.exists() && !f.isDirectory()) {
+                System.out.println("EXIST FILE................");
+            }else {
+                System.out.println("USING COPY ...............");
+                Files.copy(file.getInputStream(), this.rootLocation.resolve(file.getOriginalFilename()));
+            }
+
         } catch (Exception e) {
-            throw new RuntimeException("FAIL!");
+            logger.error(e.getMessage());
+            System.out.println("STORAGE");
+            System.out.println("Message ");
+            System.out.println(e.getMessage());
+            System.out.println("Couse ");
+            System.out.println(e.getCause());
+            throw new RuntimeException("FAIL!", e.getCause());
         }
     }
 
     public Resource loadFile(String filename) {
+        logger.info("Call loadFile: {}", filename);
         try {
             Path file = rootLocation.resolve(filename);
             Resource resource = new UrlResource(file.toUri());
@@ -45,11 +62,13 @@ public class StorageService {
     }
 
     public void deleteAll() {
+        logger.info("Call deleteAll all files in bucket");
         FileSystemUtils.deleteRecursively(rootLocation.toFile());
     }
 
     public void init() {
         try {
+            logger.info("Init container to files");
             Files.createDirectory(rootLocation);
         } catch (IOException e) {
             throw new RuntimeException("Could not initialize storage!");
